+++ src/app/query-explainer/components/input-data-section/input-data-section.component.ts
@@ -1,7 +1,6 @@
 import { ChangeDetectionStrategy, Component, Input } from "@angular/core";
 import { Observable } from "rxjs";
 import {
-    QueryExplainerComponentData,
     VerifyQueryDatasetBlockNotFoundError,
     VerifyQueryDatasetNotFoundError,
     VerifyQueryError,
@@ -10,6 +9,7 @@ import {
 import { DatasetInfo } from "src/app/interface/navigation.interface";
 import { MaybeUndefined } from "src/app/common/app.types";
 import AppValues from "src/app/common/app.values";
+import { QueryExplainerComponentData } from "../../query-explainer.component";
 
 @Component({
     selector: "app-input-data-section",
@@ -22,21 +22,6 @@ export class InputDataSectionComponent {
     @Input({ required: true }) public datasetInfoObservables$: Observable<DatasetInfo>[];
     @Input({ required: true }) inputData: QueryExplainerComponentData;
     public readonly DATE_FORMAT = AppValues.DISPLAY_FLOW_DATE_FORMAT;
-
-    public inputParamsHelper(option: string): string {
-        switch (option) {
-            case "SqlDataFusion":
-                return "SQL DataFusion";
-            case "JsonAoA":
-            case "JsonAoa":
-                return "JSON AoA";
-            case "ArrowJson":
-                return "Arrow JSON";
-            /* istanbul ignore next */
-            default:
-                return "Unknown options";
-        }
-    }
 
     public isDatasetBlockNotFoundError(error: MaybeUndefined<VerifyQueryError>, blockHash: string): boolean {
         return (