+++ src/app/query-explainer/query-explainer.service.ts
@@ -1,10 +1,15 @@
 import { HttpClient, HttpErrorResponse } from "@angular/common/http";
 import { inject, Injectable } from "@angular/core";
-import { catchError, EMPTY, Observable, of } from "rxjs";
+import { catchError, EMPTY, map, Observable, of, tap } from "rxjs";
 import { AppConfigService } from "src/app/app-config.service";
-import { QueryExplainerInputType, QueryExplainerResponse, VerifyQueryResponse } from "./query-explainer.types";
+import {
+    QueryExplainerDataJsonAosResponse,
+    QueryExplainerInputType,
+    QueryExplainerProofResponse,
+    QueryExplainerResponse,
+    VerifyQueryResponse,
+} from "./query-explainer.types";
 import { ToastrService } from "ngx-toastr";
-import { LocalStorageService } from "src/app/services/local-storage.service";
 
 @Injectable({
     providedIn: "root",
@@ -13,17 +18,47 @@ export class QueryExplainerService {
     private appConfigService = inject(AppConfigService);
     private http = inject(HttpClient);
     private toastrService = inject(ToastrService);
-    private localStorageService = inject(LocalStorageService);
+    private baseUrl: string;
 
-    public processQuery(query: string): Observable<QueryExplainerResponse> {
-        const url = new URL(`${this.appConfigService.apiServerHttpUrl}/query`);
+    constructor() {
+        this.baseUrl = `${this.appConfigService.apiServerHttpUrl}`;
+    }
+
+    public processQueryWithProof(query: string): Observable<QueryExplainerProofResponse> {
+        const url = new URL(`${this.baseUrl}/query`);
         const body: QueryExplainerInputType = {
-            query: query,
-            dataFormat: "JsonAoa",
-            schemaFormat: "ArrowJson",
+            query,
             include: ["Proof"],
         };
-        return this.http.post<QueryExplainerResponse>(url.href, body).pipe(
+        return this.http.post<QueryExplainerProofResponse>(url.href, body).pipe(
+            map((response: QueryExplainerResponse) => {
+                const cloneData = Object.assign({}, response);
+                if ("output" in cloneData) {
+                    delete cloneData.output;
+                }
+                return cloneData as QueryExplainerProofResponse;
+            }),
+            tap((res: QueryExplainerProofResponse) => {
+                if (res.subQueries?.length) {
+                    throw new Error("Unsupported subQueries array");
+                }
+            }),
+            catchError((e: HttpErrorResponse) => {
+                this.toastrService.error("", e.error as string, {
+                    disableTimeOut: "timeOut",
+                });
+                return EMPTY;
+            }),
+        );
+    }
+
+    public processQueryWithSchema(query: string): Observable<QueryExplainerDataJsonAosResponse> {
+        const url = new URL(`${this.baseUrl}/query`);
+        const body: QueryExplainerInputType = {
+            query,
+            include: ["Schema"],
+        };
+        return this.http.post<QueryExplainerDataJsonAosResponse>(url.href, body).pipe(
             catchError((e: HttpErrorResponse) => {
                 this.toastrService.error("", e.error as string, {
                     disableTimeOut: "timeOut",
@@ -34,7 +69,7 @@ export class QueryExplainerService {
     }
 
     public verifyQuery(data: QueryExplainerResponse): Observable<VerifyQueryResponse> {
-        const url = new URL(`${this.appConfigService.apiServerHttpUrl}/verify`);
+        const url = new URL(`${this.baseUrl}/verify`);
         return this.http.post<VerifyQueryResponse>(url.href, data).pipe(
             catchError((e: HttpErrorResponse) => {
                 if (e.status === 400) return of(e.error as VerifyQueryResponse);
@@ -49,18 +84,14 @@ export class QueryExplainerService {
     }
 
     public fetchCommitmentDataByUploadToken(token: string): Observable<QueryExplainerResponse> {
-        const url = new URL(`${this.appConfigService.apiServerHttpUrl}/platform/file/upload/${token}`);
-        return this.http
-            .get<QueryExplainerResponse>(url.href, {
-                headers: { Authorization: `Bearer ${this.localStorageService.accessToken}` },
-            })
-            .pipe(
-                catchError((e: HttpErrorResponse) => {
-                    this.toastrService.error("", e.error as string, {
-                        disableTimeOut: "timeOut",
-                    });
-                    return EMPTY;
-                }),
-            );
+        const url = new URL(`${this.baseUrl}/platform/file/upload/${token}`);
+        return this.http.get<QueryExplainerResponse>(url.href).pipe(
+            catchError((e: HttpErrorResponse) => {
+                this.toastrService.error("", e.error as string, {
+                    disableTimeOut: "timeOut",
+                });
+                return EMPTY;
+            }),
+        );
     }
 }